+++ CHANGELOG.md
@@ -6,7 +6,8 @@
 
 - [Feature(25762) Alteração da cor dos preços na PDP e PDC no APP](https://dev.azure.com/ARECO/Tribo%20E-Commerce/_workitems/edit/25762)
 
----
+- [2589 - Implementação do status do cashback](https://dev.azure.com/ARECO/Tribo%20E-Commerce/_git/mobile.reserva.app/pullrequest/2589)
+
 
 ### 2.1.10
 - Remove unused files